live_trigger/requirements.txt
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.1
streamlit>=1.32.0
SQLAlchemy>=2.0.0
beautifulsoup4>=4.12.0
playwright>=1.42.0
